Move AntiAim config handling code to AntiAim.cpp
