+++ Osiris/Config.cpp
@@ -9,6 +9,7 @@
 #include "Config.h"
 #include "Helpers.h"
 #include "SDK/Platform.h"
+#include "Hacks/AntiAim.h"
 #include "Hacks/Glow.h"
 
 #ifdef _WIN32
@@ -256,14 +257,6 @@ static void from_json(const json& j, Config::Backtrack& b)
     read(j, "Ignore smoke", b.ignoreSmoke);
     read(j, "Recoil based fov", b.recoilBasedFov);
     read(j, "Time limit", b.timeLimit);
-}
-
-static void from_json(const json& j, Config::AntiAim& a)
-{
-    read(j, "Enabled", a.enabled);
-    read(j, "Pitch", a.pitch);
-    read(j, "Yaw", a.yaw);
-    read(j, "Pitch angle", a.pitchAngle);
 }
 
 static void from_json(const json& j, Config::Chams::Material& m)
@@ -534,7 +527,7 @@ void Config::load(const char8_t* name, bool incremental) noexcept
     read(j, "Triggerbot Key", triggerbotHoldKey);
 
     read<value_t::object>(j, "Backtrack", backtrack);
-    read<value_t::object>(j, "Anti aim", antiAim);
+    ::AntiAim::fromJson(j["Anti aim"]);
     ::Glow::fromJson(j["Glow"]);
     read(j, "Chams", chams);
     read<value_t::object>(j, "ESP", streamProofESP);
@@ -711,14 +704,6 @@ static void to_json(json& j, const Config::Backtrack& o, const Config::Backtrack
     WRITE("Ignore smoke", ignoreSmoke);
     WRITE("Recoil based fov", recoilBasedFov);
     WRITE("Time limit", timeLimit);
-}
-
-static void to_json(json& j, const Config::AntiAim& o, const Config::AntiAim& dummy = {})
-{
-    WRITE("Enabled", enabled);
-    WRITE("Pitch", pitch);
-    WRITE("Pitch angle", pitchAngle);
-    WRITE("Yaw", yaw);
 }
 
 static void to_json(json& j, const Config::Chams::Material& o)
@@ -1013,7 +998,7 @@ void Config::save(size_t id) const noexcept
         to_json(j["Triggerbot Key"], triggerbotHoldKey, KeyBind::NONE);
 
         j["Backtrack"] = backtrack;
-        j["Anti aim"] = antiAim;
+        j["Anti aim"] = ::AntiAim::toJson();
         j["Glow"] = ::Glow::toJson();
         j["Chams"] = chams;
         j["ESP"] = streamProofESP;
