+++ Osiris/Hacks/AntiAim.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "../ConfigStructs.h"
+
 struct UserCmd;
 struct Vector;
 
@@ -11,4 +13,9 @@ namespace AntiAim
     void menuBarItem() noexcept;
     void tabItem() noexcept;
     void drawGUI(bool contentOnly) noexcept;
+
+    // Config
+    json toJson() noexcept;
+    void fromJson(const json& j) noexcept;
+    void resetConfig() noexcept;
 }
